1、修复生产者路由不提前更新问题。 2、修复消费者统计重复导致过大问题。 3、增加集群信息模块。 4、支持分集群暂停审核和监控。 5、优化耗时任务：按照集群维度并行，子任务并发。

+++ mq-cloud/src/main/java/com/sohu/tv/mq/cloud/bo/Topic.java
@@ -21,7 +21,7 @@ public class Topic {
     // id
     private long id;
     // cluster id
-    private long clusterId;
+    private int clusterId;
     // topic name
     private String name;
     // queue num
@@ -90,11 +90,11 @@ public class Topic {
         this.id = id;
     }
 
-    public long getClusterId() {
+    public int getClusterId() {
         return clusterId;
     }
 
-    public void setClusterId(long clusterId) {
+    public void setClusterId(int clusterId) {
         this.clusterId = clusterId;
     }
